Doors
Added four doors, one at entrance, two in main room, one at exit.

Three doors open just by walking into them.  The West door in the main room is opened by touching the table object in front of it.  Closes when no one is touching it.  Table is a stand in for a button.